refactor(oracle): migrate dhcp encode and decode to dhcp plugins
