+++ tools/oracle/engine/backends/wireshark/normalize.py
@@ -102,14 +102,6 @@ _PROTOCOL_LAYER_ALIASES: dict[str, str | None] = {
 _DOT11_ROOTS = frozenset({"Dot11", "link:dot11", "link:ieee80211"})
 _RADIOTAP_ROOTS = frozenset({"RadioTap", "link:radiotap"})
 
-# DHCP magic cookie (RFC 2132) prefacing the option region.
-_DHCP_MAGIC_COOKIE = 0x63825363
-# Option codes that need no length octet and carry no payload.
-_DHCP_OPTION_PAD = 0
-_DHCP_OPTION_END = 255
-# DHCP message type option (RFC 2132).
-_DHCP_OPTION_MESSAGE_TYPE = 53
-
 
 class WiresharkNormalizationUnsupported(RuntimeError):
     """Raised when tshark cannot perform the requested parser operation."""
@@ -363,8 +355,6 @@ def _normalize_protocol_fields(
         return _normalize_ipv6_routing(
             _layer_any(layers, "ipv6.routing", "ipv6_routing")
         )
-    if layer_name == "dhcp":
-        return _normalize_dhcp(_dhcp_layer(layers), source_hex=source_hex)
     if layer_name == "linux_sll":
         return _normalize_linux_sll(_layer(layers, "sll"))
     return {}
@@ -628,140 +618,11 @@ def _normalize_rsn(layer: JSONObject) -> JSONObject:
     return output
 
 
-def _normalize_dhcp(layer: JSONObject, *, source_hex: str | None = None) -> JSONObject:
-    """Normalize a tshark BOOTP/DHCP layer to the shared oracle field names.
-
-    The normalized names match the Scapy reference backend
-    (``tools/oracle/engine/backends/scapy/normalize.py``) and the libcrafter
-    decoder (``tools/oracle/adapters/src/bin/decode_vectors.rs``): ``opcode``,
-    ``hardware_type``/``hardware_length``, ``hops``, ``transaction_id``,
-    ``seconds``, integer ``flags``, the BOOTP IPv4 fields, ``magic_cookie``,
-    ``client_hardware_address`` as ``{"hex": ...}``, plus backend-neutral
-    ``options`` (``{code, payload_hex}`` TLVs), ``option_count`` and integer
-    ``message_type``. Wireshark renamed the dissector prefix from ``bootp.`` to
-    ``dhcp.`` around 3.0, so both prefixes are accepted.
-    """
-
-    output = _fields_from_aliases(
-        layer,
-        {
-            "opcode": ("dhcp.type", "bootp.type"),
-            "hardware_type": ("dhcp.hw.type", "bootp.hw.type"),
-            "hardware_length": ("dhcp.hw.len", "bootp.hw.len"),
-            "hops": ("dhcp.hops", "bootp.hops"),
-            "transaction_id": ("dhcp.id", "bootp.id"),
-            "seconds": ("dhcp.secs", "bootp.secs"),
-            "client_ip": ("dhcp.ip.client", "bootp.ip.client"),
-            "your_ip": ("dhcp.ip.your", "bootp.ip.your"),
-            "server_ip": ("dhcp.ip.server", "bootp.ip.server"),
-            "relay_ip": ("dhcp.ip.relay", "bootp.ip.relay"),
-            "magic_cookie": ("dhcp.cookie", "bootp.cookie"),
-        },
-    )
-    _parse_int_fields(
-        output,
-        "opcode",
-        "hardware_type",
-        "hardware_length",
-        "hops",
-        "transaction_id",
-        "seconds",
-    )
-    magic = _parse_int(output.get("magic_cookie"))
-    output["magic_cookie"] = magic if magic is not None else _DHCP_MAGIC_COOKIE
-
-    chaddr = _string_field(layer, "dhcp.hw.mac_addr", "bootp.hw.mac_addr", "dhcp.hw.addr")
-    if chaddr is not None:
-        output["client_hardware_address"] = {"hex": _hex_bytes(chaddr)}
-
-    output["flags"] = _dhcp_flags(layer)
-
-    options = _dhcp_options_from_source(source_hex)
-    if options is not None:
-        output["options"] = options
-        output["option_count"] = len(options)
-        for option in options:
-            if option["code"] == _DHCP_OPTION_MESSAGE_TYPE:
-                payload = bytes.fromhex(option["payload_hex"])
-                if len(payload) == 1:
-                    output["message_type"] = payload[0]
-                break
-    else:
-        message_type = _parse_int(
-            _field(layer, "dhcp.option.dhcp", "bootp.option.dhcp")
-        )
-        if message_type is not None:
-            output["message_type"] = message_type
-    return output
-
-
-def _dhcp_flags(layer: JSONObject) -> int:
-    """Return DHCP flags as the shared integer view (broadcast bit 0x8000)."""
-
-    value = _parse_int(_field(layer, "dhcp.flags", "bootp.flags"))
-    if value is not None:
-        return value
-    if _truthy_field(layer, "dhcp.flags.bc") or _truthy_field(layer, "bootp.flags.bc"):
-        return 0x8000
-    return 0
-
-
-def _dhcp_options_from_source(source_hex: str | None) -> list[JSONObject] | None:
-    """Reconstruct backend-neutral DHCP option TLVs from the raw packet bytes.
-
-    The option region begins right after the magic cookie. Parsing the raw bytes
-    (rather than tshark's typed option views) keeps the ``{code, payload_hex}``
-    list byte-identical to the Scapy and libcrafter decoders. A malformed or
-    truncated region yields ``None`` so callers fall back to tshark's own
-    message-type field.
-    """
-
-    if not source_hex:
-        return None
-    try:
-        raw = bytes.fromhex(source_hex)
-    except ValueError:
-        return None
-    cookie = _DHCP_MAGIC_COOKIE.to_bytes(4, "big")
-    marker = raw.find(cookie)
-    if marker < 0:
-        return None
-    return _decode_dhcp_option_tlvs(raw[marker + len(cookie) :])
-
-
-def _decode_dhcp_option_tlvs(raw: bytes) -> list[JSONObject] | None:
-    # Mirrors the Scapy reference parser
-    # (tools/oracle/engine/backends/scapy/normalize.py::_decode_dhcp_option_tlvs):
-    # pad/end are single-octet options with empty payloads and END does not stop
-    # parsing, so the ``option_count`` stays byte-identical across backends. The
-    # wire bytes are identical for the offline strict-byte path, so the region
-    # after the magic cookie matches the Scapy DHCP sub-layer region exactly.
-    options: list[JSONObject] = []
-    index = 0
-    length = len(raw)
-    while index < length:
-        code = raw[index]
-        index += 1
-        if code in {_DHCP_OPTION_PAD, _DHCP_OPTION_END}:
-            options.append({"code": code, "payload_hex": ""})
-            continue
-        if index >= length:
-            return None
-        option_length = raw[index]
-        index += 1
-        if index + option_length > length:
-            return None
-        payload = raw[index : index + option_length]
-        index += option_length
-        options.append({"code": code, "payload_hex": payload.hex()})
-    return options if options else None
-
-
-def _dhcp_layer(layers: JSONObject) -> JSONObject:
-    layer = layers.get("dhcp")
-    if isinstance(layer, dict):
-        return layer
-    return _layer(layers, "bootp")
+# The DHCP tshark normalizer ``_normalize_dhcp`` and its ``_dhcp_flags`` /
+# ``_dhcp_options_from_source`` / ``_decode_dhcp_option_tlvs`` / ``_dhcp_layer``
+# helpers and the DHCP option/cookie constants moved to ``protocols/dhcp.py`` and are
+# registered in ``WIRESHARK_REGISTRY`` (so ``_normalize_protocol_fields`` routes the
+# ``dhcp`` layer to the plugin's ``normalize``).
 
 
 def _normalize_linux_sll(layer: JSONObject) -> JSONObject:
